renamed and moved axios related files to src/service to ease its usage and clarify purpose
