+++ src/common/hooks/infoRequest.ts
@@ -2,8 +2,8 @@ import { KeycloakInstance } from 'keycloak-js'
 import { useKeycloak } from '@react-keycloak/ssr'
 import { QueryClient, useQuery } from 'react-query'
 
-import { endpoints } from 'common/api-endpoints'
-import { authQueryFnFactory } from 'common/rest'
+import { endpoints } from 'service/apiEndpoints'
+import { authQueryFnFactory } from 'service/restRequests'
 
 type InfoRequest = {
   id: string
