+++ src/components/contact/ContactForm.tsx
@@ -9,8 +9,8 @@ import makeStyles from '@mui/styles/makeStyles'
 import createStyles from '@mui/styles/createStyles'
 
 import { AlertStore } from 'stores/AlertStore'
-import { createContactRequest } from 'common/rest'
-import { isAxiosError, ApiErrors, matchValidator } from 'common/api-errors'
+import { createContactRequest } from 'service/restRequests'
+import { isAxiosError, ApiErrors, matchValidator } from 'service/apiErrors'
 import { ContactFormData, ContactRequestResponse, ContactRequestInput } from 'gql/contact'
 import GenericForm from 'components/common/form/GenericForm'
 import SubmitButton from 'components/common/form/SubmitButton'
