+++ src/pages/campaigns/[slug]/index.tsx
@@ -2,7 +2,7 @@ import { dehydrate, QueryClient } from 'react-query'
 import { GetServerSideProps } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import { queryFn } from 'common/rest'
+import { queryFn } from 'service/restRequests'
 import ViewCampaignPage from 'components/campaigns/ViewCampaignPage'
 
 export const getServerSideProps: GetServerSideProps = async ({ query, locale }) => {
