+++ src/service/apiClient.ts
@@ -8,10 +8,10 @@ const {
 } = getConfig()
 
 const cache = new LRU({ max: 10 })
-export const axios = Axios.create({ baseURL: `${API_URL}/api` })
+export const apiClient = Axios.create({ baseURL: `${API_URL}/api` })
 
 export const useAxios = makeUseAxios({
-  axios,
+  axios: apiClient,
   cache,
   defaultOptions: {
     ssr: true,
