+++ src/service/restRequests.ts
@@ -1,7 +1,7 @@
 import { AxiosRequestConfig, AxiosResponse } from 'axios'
 import { MutationFunction, QueryFunction } from 'react-query'
 
-import { axios } from 'common/api-client'
+import { apiClient } from 'service/apiClient'
 import {
   SupportRequestResponse,
   SupportRequestInput,
@@ -9,18 +9,18 @@ import {
 import { ContactRequestResponse, ContactRequestInput } from 'gql/contact'
 import { CampaignResponse, CampaignInput } from 'gql/campaigns'
 
-import { endpoints } from './api-endpoints'
+import { endpoints } from './apiEndpoints'
 import { CheckoutSessionInput, CheckoutSessionResponse } from 'gql/donations'
 import { CreateBeneficiaryInput, PersonResponse } from 'gql/person'
 
 export const queryFn: QueryFunction = async function ({ queryKey }) {
-  const response = await axios.get(queryKey.join('/'))
+  const response = await apiClient.get(queryKey.join('/'))
   return await response.data
 }
 
 export const queryFnFactory = <T>(config?: AxiosRequestConfig): QueryFunction<T> =>
   async function ({ queryKey }) {
-    const response = await axios.get(queryKey.join('/'), config)
+    const response = await apiClient.get(queryKey.join('/'), config)
     return await response.data
   }
 
@@ -33,7 +33,7 @@ export const createBeneficiary: MutationFunction<
   AxiosResponse<PersonResponse>,
   CreateBeneficiaryInput
 > = async (data: CreateBeneficiaryInput) => {
-  return await axios.post<CreateBeneficiaryInput, AxiosResponse<PersonResponse>>(
+  return await apiClient.post<CreateBeneficiaryInput, AxiosResponse<PersonResponse>>(
     endpoints.person.createBeneficiary.url,
     data,
   )
@@ -43,7 +43,7 @@ export const createContactRequest: MutationFunction<
   AxiosResponse<ContactRequestResponse>,
   ContactRequestInput
 > = async (data: ContactRequestInput) => {
-  return await axios.post<ContactRequestInput, AxiosResponse<ContactRequestResponse>>(
+  return await apiClient.post<ContactRequestInput, AxiosResponse<ContactRequestResponse>>(
     endpoints.support.createInfoRequest.url,
     data,
   )
@@ -53,7 +53,7 @@ export const createSupportRequest: MutationFunction<
   AxiosResponse<SupportRequestResponse>,
   SupportRequestInput
 > = async (data: SupportRequestInput) => {
-  return await axios.post<SupportRequestInput, AxiosResponse<SupportRequestResponse>>(
+  return await apiClient.post<SupportRequestInput, AxiosResponse<SupportRequestResponse>>(
     endpoints.support.createSupportRequest.url,
     data,
   )
@@ -61,7 +61,7 @@ export const createSupportRequest: MutationFunction<
 
 export const createCampaign: MutationFunction<AxiosResponse<CampaignResponse>, CampaignInput> =
   async (data: CampaignInput) => {
-    return await axios.post<CampaignInput, AxiosResponse<CampaignResponse>>(
+    return await apiClient.post<CampaignInput, AxiosResponse<CampaignResponse>>(
       endpoints.campaign.createCampaign.url,
       data,
     )
@@ -71,7 +71,7 @@ export const createCheckoutSession: MutationFunction<
   AxiosResponse<CheckoutSessionResponse>,
   CheckoutSessionInput
 > = async (data: CheckoutSessionInput) => {
-  return await axios.post<CheckoutSessionInput, AxiosResponse<CheckoutSessionResponse>>(
+  return await apiClient.post<CheckoutSessionInput, AxiosResponse<CheckoutSessionResponse>>(
     endpoints.donation.createCheckoutSession.url,
     data,
   )
